improvements for #90

+++ WickedEngine/wiRawInput.cpp
@@ -442,7 +442,7 @@ namespace wiRawInput
 	}
 	bool GetControllerState(wiInput::ControllerState* state, int index)
 	{
-		if (index < controllers.size() && controllers[index].handle && !controllers[index].is_xinput)
+		if (index < (int)controllers.size() && controllers[index].handle && !controllers[index].is_xinput)
 		{
 			if (state != nullptr)
 			{
@@ -454,7 +454,7 @@ namespace wiRawInput
 	}
 	void SetControllerFeedback(const wiInput::ControllerFeedback& data, int index)
 	{
-		if (index < controllers.size() && controllers[index].handle && !controllers[index].is_xinput)
+		if (index < (int)controllers.size() && controllers[index].handle && !controllers[index].is_xinput)
 		{
 			HANDLE hid_device = CreateFile(
 				controllers[index].name.c_str(),
